% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innocolours.R
\name{InnoColours}
\alias{InnoColours}
\title{InnoColours()}
\usage{
InnoColours(set = "standard")
}
\arguments{
\item{set}{defaults to "standard", "bright" returns a lighter tint
suitable for dark backgrounds}
}
\description{
This function returns one of two versions of the InnoRenew colour palette
for creating figures (it does not include grey or beige)
}
\examples{
Used in any manual scale that takes a colour in a ggplot() plot...
ggplot(data=data, aes(x=x, y=y, fill=group, colour=group)) +
theme_InnoLight() +
geom_col() +
scale_fill_manual(values = InnoColours())
scale_colour_manual(values = InnoColours(set="bright"))
}
\keyword{color,}
\keyword{colour,}
\keyword{colourblindness,}
\keyword{palette}
