% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_beige.R
\name{theme_InnoBeige}
\alias{theme_InnoBeige}
\title{theme_InnoBeige()}
\usage{
theme_InnoBeige()
}
\arguments{
\item{NA}{No parameters.}
}
\description{
This ggplot theme has a beige background (#FEFEF1) and should be used
with the standard colour set from this package (InnoColours() or
InnoColours(set=="standard"))
** Warning ** calling this theme will adjust the default settings for
all geom colour and fill values. To reset this you need to restart your R
session and reload the ggplot2 (or tidyverse) package.
}
\examples{
ggplot(data=data, aes(x=x, y=y)) + theme_InnoBeige() + geom_point()
}
\keyword{dark,}
\keyword{ggplot,}
\keyword{innorenew,}
\keyword{theme}
