% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dark.R
\name{theme_InnoDark}
\alias{theme_InnoDark}
\title{theme_InnoDark()}
\usage{
theme_InnoDark()
}
\arguments{
\item{NA}{No parameters.}
}
\description{
This ggplot theme has dark grey background (#4F4F4F) and should be used
with the bright colour set from this package (InnoColours(set="bright"))
** Warning ** calling this theme will adjust the default settings for
all geom colour and fill values. To reset this you need to restart your R
session and reload the ggplot2 (or tidyverse) package.
}
\examples{
ggplot(data=data, aes(x=x, y=y)) + theme_InnoDark() + geom_point()
}
\keyword{dark,}
\keyword{ggplot,}
\keyword{innorenew,}
\keyword{theme}
